requirements.txt
flask
flask_sqlalchemy
flask_login
werkzeug
python-dotenv
reportlab
google-generativeai